requirements.txt
asttokens==2.1.0
backcall==0.2.0
bcrypt==4.0.1
-e git+https://github.com/shubhamjain31/Railway_Management_System.git@f559aa51d9c3193fe562ab823b9ad242c720f153#egg=core
cryptacular==1.6.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
executing==1.2.0
greenlet==2.0.1
gunicorn==20.1.0
hupper==1.10.3
idna==3.4
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
PasteDeploy==3.0.1
pbkdf2==1.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
plaster==1.1
plaster-pastedeploy==1.0.1
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyramid==2.0
pyramid-bootstrap==4.1.0
pyramid-debugtoolbar==4.9
pyramid-exclog==1.1
pyramid-ipython==0.2
pyramid-jinja2==2.10
pyramid-layout==1.0
pyramid-mako==1.1.0
pyramid-sqlalchemy==1.6
pyramid-tm==2.5
python-decouple==3.6
repoze.lru==0.7
six==1.16.0
SQLAlchemy==1.4.44
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
stack-data==0.6.1
Tempita==0.5.2
traitlets==5.5.0
transaction==3.0.1
translationstring==1.4
venusian==3.0.0
waitress==2.1.2
wcwidth==0.2.5
WebOb==1.8.7
WTForms==3.0.1
zope.deprecation==4.4.0
zope.interface==5.5.1
zope.sqlalchemy==1.6
